Move KDF to nfc_helpers
