+++ firmware/components/nfc/nfc.h
@@ -52,7 +52,28 @@ static const char *const ERROR = "Error";
 static const uint8_t DEFAULT_KEY[6] = {0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF};
 static const uint8_t NDEF_KEY[6] = {0xD3, 0xF7, 0xD3, 0xF7, 0xD3, 0xF7};
 static const uint8_t MAD_KEY[6] = {0xA0, 0xA1, 0xA2, 0xA3, 0xA4, 0xA5};
-static const uint8_t BAMBU_SALT[16] = {0x9a, 0x75, 0x9c, 0xf2, 0xc4, 0xf7, 0xca, 0xff,0x22, 0x2c, 0xb9, 0x76, 0x9b, 0x41, 0xbc, 0x96};
+// static const uint8_t BAMBU_SALT[16] = {0x9a, 0x75, 0x9c, 0xf2, 0xc4, 0xf7, 0xca, 0xff,0x22, 0x2c, 0xb9, 0x76, 0x9b, 0x41, 0xbc, 0x96};
+// static const uint8_t KEY0[6] = {0x63, 0xe5, 0xaf, 0x2c, 0x1d, 0x75};
+// static const uint8_t KEY1[6] = {0x40, 0xd1, 0x46, 0xce, 0x6e, 0x01};
+// static const uint8_t KEY2[6] = {0x6a, 0x66, 0x95, 0x7d, 0xcc, 0x91};
+// static const uint8_t KEY3[6] = {0x15, 0xe7, 0x04, 0x1f, 0x68, 0xd9};
+// static const uint8_t KEY4[6] = {0x7e, 0xe1, 0xac, 0x7f, 0xa7, 0x5f};
+// static const uint8_t KEY5[6] = {0x55, 0xcb, 0xba, 0xd1, 0x86, 0x73};
+// static const uint8_t KEY6[6] = {0xce, 0x59, 0x01, 0xaf, 0x94, 0x16};
+// static const uint8_t KEY7[6] = {0xa2, 0x23, 0xa1, 0x93, 0xe6, 0xa3};
+// static const uint8_t KEY8[6] = {0x24, 0xf4, 0xd0, 0x22, 0xf4, 0x02};
+// static const uint8_t KEY9[6] = {0x7d, 0xf9, 0x99, 0xdd, 0x83, 0x6b};
+// static const uint8_t KEY10[6] = {0xb0, 0xda, 0xc4, 0xa4, 0x89, 0x03};
+// static const uint8_t KEY11[6] = {0xb0, 0x26, 0xab, 0x56, 0x6f, 0x11};
+// static const uint8_t KEY12[6] = {0x8b, 0x49, 0x5d, 0x5a, 0x0b, 0x44};
+// static const uint8_t KEY13[6] = {0x7e, 0xbe, 0xf1, 0xcb, 0x3e, 0x94};
+// static const uint8_t KEY14[6] = {0x46, 0x85, 0x79, 0x0c, 0x6e, 0x01};
+// static const uint8_t KEY15[6] = {0x3f, 0x00, 0x14, 0x4c, 0x7b, 0x4a};
+// static const std::array<const uint8_t*, 16> KEYS = {{
+//     KEY0, KEY1, KEY2, KEY3, KEY4, KEY5, KEY6, KEY7, 
+//     KEY8, KEY9, KEY10, KEY11, KEY12, KEY13, KEY14, KEY15
+// }};
+
 
 std::string format_uid(std::vector<uint8_t> &uid);
 std::string format_bytes(std::vector<uint8_t> &bytes);
@@ -67,10 +88,6 @@ bool mifare_classic_is_trailer_block(uint8_t block_num);
 
 uint32_t get_mifare_ultralight_buffer_size(uint32_t message_length);
 
-// std::array<std::array<uint8_t, 6>, 16> generate_keys(const std::vector<uint8_t>& uid, const uint8_t* salt, size_t salt_size);
-// std::array<std::array<uint8_t, 6>, 16> generate_keys(const std::vector<uint8_t>& uid);
-// std::vector<int> generate_keys();
-std::vector<std::vector<uint8_t>> generate_keys();
 
 class NfcTagListener {
  public:
